added reusable airtable component and refactors latest news page to use the component
