+++ src/components/News/components/LatestNews.jsx
@@ -1,123 +1,173 @@
-const posts = [
-    {
-      id: 1,
-      title: 'Boost your conversion rate',
-      href: '#',
-      description:
-        'Illo sint voluptas. Error voluptates culpa eligendi. Hic vel totam vitae illo. Non aliquid explicabo necessitatibus unde. Sed exercitationem placeat consectetur nulla deserunt vel. Iusto corrupti dicta.',
-      imageUrl:
-        'https://images.unsplash.com/photo-1496128858413-b36217c2ce36?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=3603&q=80',
-      date: 'Mar 16, 2020',
-      datetime: '2020-03-16',
-      category: { title: 'Marketing', href: '#' },
-      author: {
-        name: 'Michael Foster',
-        role: 'Co-Founder / CTO',
-        href: '#',
-        imageUrl:
-          'https://images.unsplash.com/photo-1519244703995-f4e0f30006d5?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=facearea&facepad=2&w=256&h=256&q=80',
-      },
-    },
-    {
-        id: 2,
-        title: 'Boost your conversion rate',
-        href: '#',
-        description:
-          'Illo sint voluptas. Error voluptates culpa eligendi. Hic vel totam vitae illo. Non aliquid explicabo necessitatibus unde. Sed exercitationem placeat consectetur nulla deserunt vel. Iusto corrupti dicta.',
-        imageUrl:
-          'https://images.unsplash.com/photo-1496128858413-b36217c2ce36?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=3603&q=80',
-        date: 'Mar 16, 2020',
-        datetime: '2020-03-16',
-        category: { title: 'Marketing', href: '#' },
-        author: {
-          name: 'Michael Foster',
-          role: 'Co-Founder / CTO',
-          href: '#',
-          imageUrl:
-            'https://images.unsplash.com/photo-1519244703995-f4e0f30006d5?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=facearea&facepad=2&w=256&h=256&q=80',
-        },
-      },
-      {
-        id: 3,
-        title: 'Boost your conversion rate',
-        href: '#',
-        description:
-          'Illo sint voluptas. Error voluptates culpa eligendi. Hic vel totam vitae illo. Non aliquid explicabo necessitatibus unde. Sed exercitationem placeat consectetur nulla deserunt vel. Iusto corrupti dicta.',
-        imageUrl:
-          'https://images.unsplash.com/photo-1496128858413-b36217c2ce36?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=3603&q=80',
-        date: 'Mar 16, 2020',
-        datetime: '2020-03-16',
-        category: { title: 'Marketing', href: '#' },
-        author: {
-          name: 'Michael Foster',
-          role: 'Co-Founder / CTO',
-          href: '#',
-          imageUrl:
-            'https://images.unsplash.com/photo-1519244703995-f4e0f30006d5?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=facearea&facepad=2&w=256&h=256&q=80',
-        },
-      },
-  ]
-  
-  export default function LatestNews() {
-    return (
-      <div className="bg-white py-12 sm:py-16">
-        <div className="mx-auto max-w-7xl px-6 lg:px-8">
-          <div className="mx-auto max-w-2xl text-center">
-            <h2 className="text-3xl font-bold tracking-tight text-gray-900 sm:text-4xl">Latest News</h2>
-            {/* <p className="mt-2 text-lg leading-8 text-gray-600">
-              Learn how to grow your business with our expert advice.
-            </p> */}
-          </div>
-          <div className="mx-auto mt-16 grid max-w-2xl grid-cols-1 gap-x-8 gap-y-20 lg:mx-0 lg:max-w-none lg:grid-cols-3">
-            {posts.map((post) => (
-              <article key={post.id} className="flex flex-col items-start justify-between">
-                <div className="relative w-full">
-                  <img
-                    alt=""
-                    src={post.imageUrl}
-                    className="aspect-[16/9] w-full rounded-2xl bg-gray-100 object-cover sm:aspect-[2/1] lg:aspect-[3/2]"
-                  />
-                  <div className="absolute inset-0 rounded-2xl ring-1 ring-inset ring-gray-900/10" />
-                </div>
-                <div className="max-w-xl">
-                  {/* <div className="mt-8 flex items-center gap-x-4 text-xs">
-                    <time dateTime={post.datetime} className="text-gray-500">
-                      {post.date}
-                    </time>
-                    <a
-                      href={post.category.href}
-                      className="relative z-10 rounded-full bg-gray-50 px-3 py-1.5 font-medium text-gray-600 hover:bg-gray-100"
-                    >
-                      {post.category.title}
-                    </a>
-                  </div> */}
-                  <div className="group relative">
-                    <h3 className="mt-3 text-lg font-semibold leading-6 text-gray-900 group-hover:text-gray-600">
-                      <a href={post.href}>
-                        <span className="absolute inset-0" />
-                        {post.title}
-                      </a>
-                    </h3>
-                    <p className="mt-5 line-clamp-3 text-sm leading-6 text-gray-600">{post.description}</p>
-                  </div>
-                  {/* <div className="relative mt-8 flex items-center gap-x-4">
-                    <img alt="" src={post.author.imageUrl} className="h-10 w-10 rounded-full bg-gray-100" />
-                    <div className="text-sm leading-6">
-                      <p className="font-semibold text-gray-900">
-                        <a href={post.author.href}>
-                          <span className="absolute inset-0" />
-                          {post.author.name}
-                        </a>
-                      </p>
-                      <p className="text-gray-600">{post.author.role}</p>
-                    </div>
-                  </div> */}
-                </div>
-              </article>
-            ))}
-          </div>
+// const posts = [
+//     {
+//       id: 1,
+//       title: 'Boost your conversion rate',
+//       href: '#',
+//       description:
+//         'Illo sint voluptas. Error voluptates culpa eligendi. Hic vel totam vitae illo. Non aliquid explicabo necessitatibus unde. Sed exercitationem placeat consectetur nulla deserunt vel. Iusto corrupti dicta.',
+//       imageUrl:
+//         'https://images.unsplash.com/photo-1496128858413-b36217c2ce36?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=3603&q=80',
+//       date: 'Mar 16, 2020',
+//       datetime: '2020-03-16',
+//       category: { title: 'Marketing', href: '#' },
+//       author: {
+//         name: 'Michael Foster',
+//         role: 'Co-Founder / CTO',
+//         href: '#',
+//         imageUrl:
+//           'https://images.unsplash.com/photo-1519244703995-f4e0f30006d5?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=facearea&facepad=2&w=256&h=256&q=80',
+//       },
+//     },
+//     {
+//         id: 2,
+//         title: 'Boost your conversion rate',
+//         href: '#',
+//         description:
+//           'Illo sint voluptas. Error voluptates culpa eligendi. Hic vel totam vitae illo. Non aliquid explicabo necessitatibus unde. Sed exercitationem placeat consectetur nulla deserunt vel. Iusto corrupti dicta.',
+//         imageUrl:
+//           'https://images.unsplash.com/photo-1496128858413-b36217c2ce36?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=3603&q=80',
+//         date: 'Mar 16, 2020',
+//         datetime: '2020-03-16',
+//         category: { title: 'Marketing', href: '#' },
+//         author: {
+//           name: 'Michael Foster',
+//           role: 'Co-Founder / CTO',
+//           href: '#',
+//           imageUrl:
+//             'https://images.unsplash.com/photo-1519244703995-f4e0f30006d5?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=facearea&facepad=2&w=256&h=256&q=80',
+//         },
+//       },
+//       {
+//         id: 3,
+//         title: 'Boost your conversion rate',
+//         href: '#',
+//         description:
+//           'Illo sint voluptas. Error voluptates culpa eligendi. Hic vel totam vitae illo. Non aliquid explicabo necessitatibus unde. Sed exercitationem placeat consectetur nulla deserunt vel. Iusto corrupti dicta.',
+//         imageUrl:
+//           'https://images.unsplash.com/photo-1496128858413-b36217c2ce36?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=3603&q=80',
+//         date: 'Mar 16, 2020',
+//         datetime: '2020-03-16',
+//         category: { title: 'Marketing', href: '#' },
+//         author: {
+//           name: 'Michael Foster',
+//           role: 'Co-Founder / CTO',
+//           href: '#',
+//           imageUrl:
+//             'https://images.unsplash.com/photo-1519244703995-f4e0f30006d5?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=facearea&facepad=2&w=256&h=256&q=80',
+//         },
+//       },
+//   ]
+
+//   export default function LatestNews() {
+//     return (
+//       <div className="bg-white py-12 sm:py-16">
+//         <div className="mx-auto max-w-7xl px-6 lg:px-8">
+//           <div className="mx-auto max-w-2xl text-center">
+//             <h2 className="text-3xl font-bold tracking-tight text-gray-900 sm:text-4xl">Latest News</h2>
+//             {/* <p className="mt-2 text-lg leading-8 text-gray-600">
+//               Learn how to grow your business with our expert advice.
+//             </p> */}
+//           </div>
+//           <div className="mx-auto mt-16 grid max-w-2xl grid-cols-1 gap-x-8 gap-y-20 lg:mx-0 lg:max-w-none lg:grid-cols-3">
+//             {posts.map((post) => (
+//               <article key={post.id} className="flex flex-col items-start justify-between">
+//                 <div className="relative w-full">
+//                   <img
+//                     alt=""
+//                     src={post.imageUrl}
+//                     className="aspect-[16/9] w-full rounded-2xl bg-gray-100 object-cover sm:aspect-[2/1] lg:aspect-[3/2]"
+//                   />
+//                   <div className="absolute inset-0 rounded-2xl ring-1 ring-inset ring-gray-900/10" />
+//                 </div>
+//                 <div className="max-w-xl">
+//                   {/* <div className="mt-8 flex items-center gap-x-4 text-xs">
+//                     <time dateTime={post.datetime} className="text-gray-500">
+//                       {post.date}
+//                     </time>
+//                     <a
+//                       href={post.category.href}
+//                       className="relative z-10 rounded-full bg-gray-50 px-3 py-1.5 font-medium text-gray-600 hover:bg-gray-100"
+//                     >
+//                       {post.category.title}
+//                     </a>
+//                   </div> */}
+//                   <div className="group relative">
+//                     <h3 className="mt-3 text-lg font-semibold leading-6 text-gray-900 group-hover:text-gray-600">
+//                       <a href={post.href}>
+//                         <span className="absolute inset-0" />
+//                         {post.title}
+//                       </a>
+//                     </h3>
+//                     <p className="mt-5 line-clamp-3 text-sm leading-6 text-gray-600">{post.description}</p>
+//                   </div>
+//                   {/* <div className="relative mt-8 flex items-center gap-x-4">
+//                     <img alt="" src={post.author.imageUrl} className="h-10 w-10 rounded-full bg-gray-100" />
+//                     <div className="text-sm leading-6">
+//                       <p className="font-semibold text-gray-900">
+//                         <a href={post.author.href}>
+//                           <span className="absolute inset-0" />
+//                           {post.author.name}
+//                         </a>
+//                       </p>
+//                       <p className="text-gray-600">{post.author.role}</p>
+//                     </div>
+//                   </div> */}
+//                 </div>
+//               </article>
+//             ))}
+//           </div>
+//         </div>
+//       </div>
+//     )
+//   }
+
+import React from "react";
+import Airtable from "../../common/Airtable/Airtable";
+
+export default function LatestNews() {
+
+  const renderNewsItem = (post) => ( 
+    <article key={post.id} className="flex flex-col items-start justify-between">
+      <div className="relative w-full">
+        <img
+          alt=""
+          src={post.fields.imageUrl}
+          className="aspect-[16/9] w-full rounded-2xl bg-gray-100 object-cover sm:aspect-[2/1] lg:aspect-[3/2]"
+        />
+        <div className="absolute inset-0 rounded-2xl ring-1 ring-inset ring-gray-900/10" />
+      </div>
+      <div className="max-w-xl">
+        <div className="group   
+ relative">
+          <h3 className="mt-3 text-lg font-semibold leading-6 text-gray-900 group-hover:text-gray-600">
+            <a href={post.fields.href}>   
+
+              <span className="absolute inset-0" />
+              {post.fields.title}
+            </a>
+          </h3>
+          <p className="mt-5 line-clamp-3 text-sm leading-6 text-gray-600">{post.fields.description}</p>
         </div>
       </div>
-    )
-  }
-  
+    </article>
+  );
+    
+  return (
+    <div className="bg-white py-12 sm:py-16">
+      <div className="mx-auto max-w-7xl px-6 lg:px-8">
+        <div className="mx-auto max-w-2xl text-center">
+          <h2 className="text-3xl   
+ font-bold tracking-tight text-gray-900 sm:text-4xl">Latest   
+ News</h2>
+        </div>
+        <div className="mx-auto mt-16 grid max-w-2xl grid-cols-1 gap-x-8 gap-y-20 lg:mx-0 lg:max-w-none lg:grid-cols-3">
+         <Airtable
+          tableName="News"
+          view="Grid view"
+          renderItem={renderNewsItem}
+        />
+        </div>
+      </div>
+    </div>
+  );
+}
